refactor: peel self-healing constants/branch + dashboard API client/health
Wave 4 of package organization:
- self-healing-constants.ts + self-healing-branch.ts (isBranchAheadOfBase)
- dashboard app/api/client.ts (api wrapper) and health.ts (health/engine/updates)
- Keep public re-exports on self-healing.ts and legacy.ts
