+++ scripts/line-count-baseline.json
@@ -17,7 +17,7 @@
   "packages/core/src/postgres/sqlite-migrator.ts": 2068,
   "packages/core/src/store.ts": 2631,
   "packages/core/src/types.ts": 7336,
-  "packages/dashboard/app/api/legacy.ts": 11627,
+  "packages/dashboard/app/api/legacy.ts": 11557,
   "packages/dashboard/app/components/AgentDetailView.tsx": 5552,
   "packages/dashboard/app/components/AgentsView.tsx": 2254,
   "packages/dashboard/app/components/ChatView.tsx": 3850,
@@ -96,7 +96,7 @@
   "packages/engine/src/pi.ts": 2799,
   "packages/engine/src/project-engine.ts": 5051,
   "packages/engine/src/scheduler.ts": 3370,
-  "packages/engine/src/self-healing.ts": 12183,
+  "packages/engine/src/self-healing.ts": 12400,
   "packages/engine/src/triage.ts": 3133,
   "packages/pi-claude-cli/src/__tests__/provider.test.ts": 2028,
   "plugins/fusion-plugin-roadmap/src/dashboard/RoadmapsView.tsx": 2583,
